requirements.txt
ultralytics
opencv-python
torch
torchvision
Pillow
fastapi
uvicorn
numpy
scikit-learn
uvicorn